Improve reading and printing rationals

+++ src/global.h
@@ -22,7 +22,7 @@ enum ErrorCode {Success, Error};
 typedef enum ErrorCode ErrorCode_t;
 
 void error(const char* s, ...);
-void debug(int, const char* s, ...);
+void debug(int8_t, const char* s, ...);
 
 INDEX bytes_to_index(uint8_t*, uint8_t);
 void index_to_bytes(INDEX, uint8_t*, uint8_t);
